WIP Needs Testing : Asynchronous loading of relations
For get, attach, and detach, we need to asynchronously fetch the related data if it hasn't been retrieved yet. This would be for the case where the first connection of related resources is retrieved. But following connections aren't. Like getting an Author with all their Books. But then each Book has a Publisher or list of Contributors. So the Publisher and Contributor would need to be fetched.